requirements.txt
tensorflow
tf_keras
opencv-python
Pillow
numpy
